requirements.txt
aiogram==2.25.1
telethon==1.28.5
python-docx==0.8.11
python-dotenv==1.0.0